+++ lib/commercetools-api/src/Models/ProductTailoring/ProductTailoringAttributeModel.php
@@ -1,0 +1,112 @@
+<?php
+
+declare(strict_types=1);
+/**
+ * This file has been auto generated
+ * Do not change it.
+ */
+
+namespace Commercetools\Api\Models\ProductTailoring;
+
+use Commercetools\Base\DateTimeImmutableCollection;
+use Commercetools\Base\JsonObject;
+use Commercetools\Base\JsonObjectModel;
+use Commercetools\Base\MapperFactory;
+use stdClass;
+
+/**
+ * @internal
+ */
+final class ProductTailoringAttributeModel extends JsonObjectModel implements ProductTailoringAttribute
+{
+    /**
+     *
+     * @var ?string
+     */
+    protected $name;
+
+    /**
+     *
+     * @var ?mixed
+     */
+    protected $value;
+
+
+    /**
+     * @psalm-suppress MissingParamType
+     */
+    public function __construct(
+        ?string $name = null,
+        $value = null
+    ) {
+        $this->name = $name;
+        $this->value = $value;
+    }
+
+    /**
+     * <p>Name of the Attribute.</p>
+     *
+     *
+     * @return null|string
+     */
+    public function getName()
+    {
+        if (is_null($this->name)) {
+            /** @psalm-var ?string $data */
+            $data = $this->raw(self::FIELD_NAME);
+            if (is_null($data)) {
+                return null;
+            }
+            $this->name = (string) $data;
+        }
+
+        return $this->name;
+    }
+
+    /**
+     * <p>The <a href="ctp:api:type:AttributeType">AttributeType</a> determines the format of the Attribute <code>value</code> to be provided:</p>
+     * <ul>
+     * <li>For <a href="ctp:api:type:AttributeEnumType">Enum Type</a> and <a href="ctp:api:type:AttributeLocalizedEnumType">Localized Enum Type</a>,
+     * use the <code>key</code> of the <a href="ctp:api:type:AttributePlainEnumValue">Plain Enum Value</a> or <a href="ctp:api:type:AttributeLocalizedEnumValue">Localized Enum Value</a> objects,
+     * or the complete objects as <code>value</code>.</li>
+     * <li>For <a href="ctp:api:type:AttributeLocalizableTextType">Localizable Text Type</a>, use the <a href="ctp:api:type:LocalizedString">LocalizedString</a> object as <code>value</code>.</li>
+     * <li>For <a href="ctp:api:type:AttributeMoneyType">Money Type</a> Attributes, use the <a href="ctp:api:type:Money">Money</a> object as <code>value</code>.</li>
+     * <li>For <a href="ctp:api:type:AttributeSetType">Set Type</a> Attributes, use the entire <code>set</code> object  as <code>value</code>.</li>
+     * <li>For <a href="ctp:api:type:AttributeReferenceType">Reference Type</a> Attributes, use the <a href="ctp:api:type:Reference">Reference</a> object as <code>value</code>.</li>
+     * </ul>
+     * <p>Tailoring of <a href="ctp:api:type:AttributeNestedType">Nested Type</a> Attributes is not supported.</p>
+     *
+     *
+     * @return null|mixed
+     */
+    public function getValue()
+    {
+        if (is_null($this->value)) {
+            /** @psalm-var mixed $data */
+            $data = $this->raw(self::FIELD_VALUE);
+            if (is_null($data)) {
+                return null;
+            }
+            $this->value = $data;
+        }
+
+        return $this->value;
+    }
+
+
+    /**
+     * @param ?string $name
+     */
+    public function setName(?string $name): void
+    {
+        $this->name = $name;
+    }
+
+    /**
+     * @param mixed $value
+     */
+    public function setValue($value): void
+    {
+        $this->value = $value;
+    }
+}