+++ lib/commercetools-api/src/Models/ProductTailoring/ProductVariantTailoring.php
@@ -18,6 +18,7 @@ interface ProductVariantTailoring extends JsonObject
     public const FIELD_ID = 'id';
     public const FIELD_IMAGES = 'images';
     public const FIELD_ASSETS = 'assets';
+    public const FIELD_ATTRIBUTES = 'attributes';
 
     /**
      * <p>The <code>id</code> of the tailored <a href="ctp:api:type:ProductVariant">ProductVariant</a>.</p>
@@ -46,6 +47,19 @@ interface ProductVariantTailoring extends JsonObject
     public function getAssets();
 
     /**
+     * <p>Attributes of the tailored Product Variant.
+     * If present, these Attributes are selectively merged into the <code>attributes</code> of the corresponding <a href="ctp:api:type:ProductVariant">ProductVariant</a>:</p>
+     * <ul>
+     * <li>If the ProductVariant contains an Attribute with the same <code>name</code>, its <code>value</code> is overwritten,</li>
+     * <li>otherwise the Attribute and its value are added to the ProductVariant.</li>
+     * </ul>
+     *
+
+     * @return null|ProductTailoringAttributeCollection
+     */
+    public function getAttributes();
+
+    /**
      * @param ?int $id
      */
     public function setId(?int $id): void;
@@ -59,4 +73,9 @@ interface ProductVariantTailoring extends JsonObject
      * @param ?AssetCollection $assets
      */
     public function setAssets(?AssetCollection $assets): void;
+
+    /**
+     * @param ?ProductTailoringAttributeCollection $attributes
+     */
+    public function setAttributes(?ProductTailoringAttributeCollection $attributes): void;
 }