+++ changes.md
@@ -6,6 +6,9 @@
 - added property `staged` to type `ProductVariantDeletedMessage`
 - added property `staged` to type `ProductVariantDeletedMessagePayload`
 - added property `warnings` to type `ProductTailoring`
+- added property `attributes` to type `ProductVariantTailoring`
+- added property `attributes` to type `ProductVariantTailoringDraft`
+- added property `attributes` to type `ProductTailoringAddVariantAction`
 - added property `warnings` to type `Product`
 </details>
 
@@ -13,7 +16,18 @@
 <details>
 <summary>Added Type(s)</summary>
 
+- added type `ProductTailoringAttribute`
+- added type `ProductTailoringSetAttributeAction`
+- added type `ProductTailoringSetAttributeInAllVariantsAction`
 - added type `ImageProcessingOngoingWarning`
 - added type `WarningObject`
+</details>
+
+
+<details>
+<summary>Added Enum(s)</summary>
+
+- added enum `product-tailoring` to type `MessageSubscriptionResourceTypeId`
+- added enum `product-tailoring` to type `ResourceTypeId`
 </details>